Move channel header to collapsible app bar
