--- app/src/main/res/layout/channel_header.xml
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="?attr/contrast_background_color">
-
-    <RelativeLayout
-        android:id="@+id/channel_metadata"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ImageView
-            android:id="@+id/channel_banner_image"
-            android:layout_width="match_parent"
-            android:layout_height="70dp"
-            android:background="@android:color/black"
-            android:fitsSystemWindows="true"
-            android:scaleType="fitCenter"
-            android:src="@drawable/placeholder_channel_banner"
-            tools:ignore="ContentDescription" />
-
-        <FrameLayout
-            android:id="@+id/avatars_layout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:layout_marginTop="50dp">
-
-            <com.google.android.material.imageview.ShapeableImageView
-                android:id="@+id/channel_avatar_view"
-                android:layout_width="@dimen/channel_avatar_size"
-                android:layout_height="@dimen/channel_avatar_size"
-                android:padding="1dp"
-                android:src="@drawable/placeholder_person"
-                app:shapeAppearance="@style/CircularImageView"
-                app:strokeColor="#ffffff"
-                app:strokeWidth="2dp" />
-
-            <com.google.android.material.imageview.ShapeableImageView
-                android:id="@+id/sub_channel_avatar_view"
-                android:layout_width="@dimen/sub_channel_avatar_size"
-                android:layout_height="@dimen/sub_channel_avatar_size"
-                android:layout_gravity="bottom|right"
-                android:padding="1dp"
-                android:src="@drawable/placeholder_person"
-                android:visibility="gone"
-                app:shapeAppearance="@style/CircularImageView"
-                app:strokeColor="#ffffff"
-                app:strokeWidth="2dp"
-                tools:ignore="RtlHardcoded"
-                tools:visibility="visible" />
-        </FrameLayout>
-
-        <org.schabi.newpipe.views.NewPipeTextView
-            android:id="@+id/channel_title_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/channel_banner_image"
-            android:layout_marginLeft="8dp"
-            android:layout_marginTop="6dp"
-            android:layout_marginRight="8dp"
-            android:layout_toLeftOf="@id/channel_subscribe_button"
-            android:layout_toRightOf="@id/avatars_layout"
-            android:ellipsize="end"
-            android:lines="1"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:textSize="@dimen/video_item_detail_title_text_size"
-            tools:ignore="RtlHardcoded"
-            tools:text="Lorem ipsum dolor" />
-
-        <org.schabi.newpipe.views.NewPipeTextView
-            android:id="@+id/sub_channel_title_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/channel_title_view"
-            android:layout_alignLeft="@id/channel_title_view"
-            android:layout_alignRight="@id/channel_title_view"
-            android:ellipsize="end"
-            android:gravity="center|left"
-            android:lines="1"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:textSize="12dp"
-            tools:ignore="RtlHardcoded"
-            tools:layout_below="@id/channel_title_view"
-            tools:text="Lorem ipsum dolor" />
-
-        <org.schabi.newpipe.views.NewPipeTextView
-            android:id="@+id/channel_subscriber_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/sub_channel_title_view"
-            android:layout_alignLeft="@id/channel_title_view"
-            android:layout_alignRight="@id/channel_title_view"
-            android:ellipsize="end"
-            android:maxLines="2"
-            android:textSize="@dimen/channel_subscribers_text_size"
-            android:visibility="gone"
-            tools:ignore="RtlHardcoded"
-            tools:text="123,141,411 subscribers"
-            tools:visibility="visible" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/channel_subscribe_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/channel_banner_image"
-            android:layout_alignParentRight="true"
-            android:layout_gravity="center_vertical|right"
-            android:layout_marginRight="2dp"
-            android:text="@string/subscribe_button_title"
-            android:textSize="@dimen/channel_rss_title_size"
-            android:theme="@style/ServiceColoredButton"
-            android:visibility="gone"
-            tools:ignore="RtlHardcoded"
-            tools:visibility="visible" />
-    </RelativeLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/channel_metadata">
-
-        <include
-            android:id="@+id/playlist_control"
-            layout="@layout/playlist_control" />
-
-    </LinearLayout>
-
-</RelativeLayout>
